+++ src/Symfony/Component/Console/CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
  * added option to run suggested command if command is not found and only 1 alternative is available
  * added option to modify console output and print multiple modifiable sections
+ * added option to use an environment variable to control color formatting
 
 4.0.0
 -----